improved arrangement, removed target from git
