--- .idea/misc.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-    <component name="ExternalStorageConfigurationManager" enabled="true" />
-    <component name="MavenProjectsManager">
-        <option name="originalFiles">
-            <list>
-                <option value="$PROJECT_DIR$/pom.xml" />
-            </list>
-        </option>
-        <option name="ignoredFiles">
-            <set>
-                <option value="$PROJECT_DIR$/src/pom.xml" />
-            </set>
-        </option>
-    </component>
-    <component name="ProjectRootManager" version="2" languageLevel="JDK_24" default="true" project-jdk-name="openjdk-24" project-jdk-type="JavaSDK">
-        <output url="file://$PROJECT_DIR$/out" />
-    </component>
-</project>
