+++ src/main/java/listeners/ContextListener.java
@@ -20,8 +20,8 @@ public class ContextListener implements ServletContextListener {
     public void contextInitialized(ServletContextEvent event) {
         DatabaseConnector dbc = null;
         try {
-            dbc = DatabaseConnector.getInstance("jdbc:mysql://localhost:3306/g",
-                    "root", "gzobava2005");
+            dbc = DatabaseConnector.getInstance("jdbc:mysql://localhost:3306/metro",
+                    "icosahedron", "Loko_kina1");
         } catch (ClassNotFoundException e) {
             throw new RuntimeException(e);
         }
